First draft of status page for managed hosts
While maintaining managed hosts in Foreman sometimes the associated VM in WMware may have been deleted or changed. Therefore a dashboard page is added to list all managed hosts that for which no VM can be found in VMware by checking the uuid.

The new dashboard page can be accessed via the Hosts menu. Unfortunately due to the way the vms are fetched the page may take a long time to load, further testing is needed. It is also a good candidate to either group by compute resource & fetching the data asynchronously via JavaScript.